Extract strings stream utilities into 'tests.utils' module
